+++ Source/UEProjectIntelligence/Public/UEProjectIntelligence.h
@@ -22,10 +22,8 @@ private:
 	FText GetDashboardStatusText() const;
 	FText GetLastActionText() const;
 	FReply RunMetadataScan();
-	FReply StartLiveWorker();
 	FReply WriteIncrementalEventsSnapshot();
 	FReply OpenSavedFolder() const;
-	FReply OpenWebUi() const;
 
 	FString LastActionMessage;
 };